Refactor code-structure
